Implement Extensions Store for feature management
This commit introduces a new `ExtensionsStore` component that allows users to toggle specific features (Beacon, RingerSong, Map, etc.) on and off.
- Adds `web/src/ExtensionsStore.jsx` with search, categories, and Quick Setup presets.
- Refactors `web/src/App.jsx` to use `ExtensionsStore` and conditionally render Sidebar/Home items based on `remoteSettings`.
- Updates Playwright tests to align with the new UI headings.
- Verification: Validated with automated tests and visual inspection via Playwright screenshot.
